Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,7 +2,7 @@
 
 int main() {
     int radius;
-    int pi = 3.14; 
+    float pi = 3.14; 
     printf("Area: %.2f", pi*radius^2);
     return 0;
 }
